consolidating service tests
